requirements.txt
streamlit==1.44.1 
streamlit-option-menu==0.4.0 
python-dotenv==1.1.0 
langchain==0.3.24  
langchain-community==0.3.22 
langchain-google-genai==2.0.10
google-genai==1.12.1
chromadb==0.6.3
PyPDF2==3.0.1 
python-docx==1.1.0
langchain-chroma==0.2.3
nltk==3.9.1
rank-bm25==0.2.2